+++ mediation/Vungle/source/plugin/Assets/GoogleMobileAds/Api/Mediation/Vungle/Vungle.cs
@@ -30,6 +30,11 @@ namespace GoogleMobileAds.Api.Mediation.Vungle
     {
         public static readonly IVungleClient client = GetVungleClient();
 
+        public static void UpdateConsentStatus(VungleConsent consentStatus)
+        {
+            client.UpdateConsentStatus(consentStatus);
+        }
+
         public static void UpdateConsentStatus(VungleConsent consentStatus,
                                                String consentMessageVersion)
         {